moving data file into a different folder
