+++ tornjak-frontend/src/components/agent-list.js
@@ -2,7 +2,7 @@ import React, { Component } from 'react';
 import { connect } from 'react-redux';
 import IsManager from './is_manager';
 import Table from "tables/agents-list-table";
-import { selectors, workloadSelectors} from '../res/data';
+import { selectors, workloadSelectors} from '../data/data';
 import TornjakApi from './tornjak-api-helpers';
 import {
   serverSelectedFunc,
